pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = "FastAPI project with Poetry"
authors = ["mdesantes <marco.desantes@globalincubator.com>"]


[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
sentry-sdk = "^1.29.0"
sqlalchemy = "^2.0.0"
alembic = "^1.11.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.0"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.4.0"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--verbose -s --cov=app --cov-report=term-missing"